Se crea pantalla de perfil del productor y se renombran componentes de producto
