+++ index/static/organico-app/src/app/app.module.ts
@@ -26,7 +26,8 @@ import { RespuestaForoComponent } from './administrador/foro/respuestaForo.compo
 import { ActivarCorreoComponent } from "./productor/activar-correo/activar-correo.component";
 import { CrearServicioComponent } from './cooperativa/servicio/crearServicio.component';
 import { ListadoServicioComponent } from './cooperativa/servicio/listadoServicio.component';
-import { CrearProductoComponent } from './cooperativa/Productos/crearProducto.component';
+import { CrearProductoComponent } from './cooperativa/productos/crearproducto.component';
+import { ProductorProfileComponent } from './productor/productor-profile/productor-profile.component';
 
 @NgModule({
   declarations: [
@@ -47,8 +48,8 @@ import { CrearProductoComponent } from './cooperativa/Productos/crearProducto.co
     RegistroAdminComponent,
     ActivarCorreoComponent,
     CrearServicioComponent,
-    CrearProductoComponent,    
-    ListadoServicioComponent
+    CrearProductoComponent,
+    ListadoServicioComponent, ProductorProfileComponent
   ],
   imports: [
     BrowserModule,
@@ -84,7 +85,7 @@ import { CrearProductoComponent } from './cooperativa/Productos/crearProducto.co
       {
         path: 'cooperativa/Productos/crearProducto',
         component: CrearProductoComponent
-      },                 
+      },
       {
         path: 'cooperativa/detalle/:id',
         component: CooperativaDetalleComponent
@@ -140,6 +141,10 @@ import { CrearProductoComponent } from './cooperativa/Productos/crearProducto.co
       {
         path: 'cooperativa/servicio/listadoServicios/:id',
         component: ListadoServicioComponent
+      },
+      {
+        path: 'profile',
+        component: ProductorProfileComponent
       }
     ])
   ],
