--- index/static/organico-app/src/app/cooperativa/Productos/CrearProducto.Component.html
@@ -1,67 +1,0 @@
-<!--Contact Section-->
-<section class="contact-section">
-  <div class="auto-container">
-      <!--Section Title-->
-        <h1>Crear Productos</h1>
-        <div class="contact-form default-form">
-
-           <form #formProductos="ngForm" (ngSubmit)="guardarProductos(formProductos)">
-                <div class="row clearfix">
-                
-                    <div class="form-group col-md-6 col-sm-6 col-xs-12">
-                        <input #nombre="ngModel" [ngClass]="{'error':!nombre.valid && envioFormProductos}"
-                               type="text" name="nombre" value=""
-                               placeholder="* Nombre del Producto [Máx 50 caracteres]" [(ngModel)]="producto.nombre"
-                               required maxlength="50" id="nombreProducto">
-                    </div>
-                    
-                    <div class="form-group col-md-6 col-sm-6 col-xs-12">
-                        <input #valor="ngModel" [ngClass]="{'error':!valor.valid && envioFormProductos}"
-                               type="text" name="valor" value=""
-                               placeholder="* Valor del Producto [Máx 6 números]" [(ngModel)]="producto.valor"
-                               required maxlength="6" pattern="[0-9-]*" id="valorProducto">                      
-                    </div>
-
-
-                    <div class="form-group col-md-12 col-sm-12 col-xs-12">
-                      <select #cooperativa="ngModel" id="cooperativa"
-                              name="cooperativa" placeholder="* Cooperativa" [(ngModel)]="producto.cooperativa" required >
-                        <option value="-1">-- Cooperativa que produce --</option>
-                        <option *ngFor="let cooperativa of cooperativas" value="{{cooperativa.id}}">{{cooperativa.nombre}}</option>
-                      </select>
-                    </div>
-
-
-
-                    <div class="form-group col-md-12 col-sm-12 col-xs-12">
-                        <textarea #descripcion="ngModel" [ngClass]="{'error':!descripcion.valid && envioFormProductos}"
-                                  name="descripcion" placeholder="* Descripción [Máx 550 caracteres]"
-                                  [(ngModel)]="producto.descripcion" maxlength="550" id="descripcionProducto"></textarea>
-                    </div>
-
-                    <div class="form-group col-md-12 col-sm-12 col-xs-12">
-                      <label> Imagen*: </label>
-                      <input type="file" id="fotoProducto" #entradaFoto (change)="loadFoto(entradaFoto)">
-                      <canvas max-width="200" max-height="200" id="photoPreview"></canvas>
-                    </div>
-
-                    <div class="form-group col-md-12 col-sm-12 col-xs-12 text-left" >
-                      <span *ngIf="!formProductos.valid && !envioFormProductos"> * Campos obligatorios</span>
-                      <br>
-                      <label *ngIf="!formProductos.valid && envioFormProductos" class="error">* Existen campos vacios que son obligatorios o existen campos invalidos.</label>
-                      <br>
-                      <span *ngIf="exitoso==true">Producto subido exitosamente</span>
-                    </div>
-
-                    <div class="form-group col-md-12 col-sm-12 col-xs-12">
-                        <div class="text-center"><button id="guardar"  type="submit" class="theme-btn btn-style-two ui-state-disabled">Guardar Producto</button>
-                        <button type="button" class="theme-btn btn-style-two ui-state-disabled" onclick="window.history.go(-1)">Cancelar</button></div>
-
-                    </div>
-
-                </div>
-
-            </form>
-        </div>
-    </div>
-</section>
