--- index/static/organico-app/src/app/cooperativa/Productos/listadoProducto.component.css
@@ -1,12 +1,0 @@
-@import "~@angular/material/prebuilt-themes/indigo-pink.css";
-
-/*.example-container {
-  display: flex;
-  flex-direction: column;
-  max-height: 500px;
-  min-width: 300px;
-}
-.mat-table {
-  overflow: auto;
-  max-height: 500px;
-}*/
