+++ index/static/organico-app/src/app/productor/productor-profile/productor-profile.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-productor-profile',
+  templateUrl: './productor-profile.component.html',
+  styleUrls: ['./productor-profile.component.css']
+})
+export class ProductorProfileComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
