+++ index/templates/base.html
@@ -81,7 +81,7 @@
                                         <ul>
                                             <li><a href="{% url 'crear_cooperativa' %}">Crear Cooperativas</a></li>
                                             <li><a href="{% url 'listado_cooperativa' %}">Listado de Cooperativas</a>
-                                            <li><a href="{% url 'crear_producto' %}">Crear Productos</a></li>   
+                                            <li><a href="{% url 'crear_producto' %}">Crear Productos</a></li>
                                             </li>
                                             <!--<li><a href="/cooperativa/mapaCooperativa/">Ver mapa de cooperativas</a></li>-->
                                         </ul>
@@ -107,6 +107,7 @@
                                     {% if user.is_authenticated %}
                                         <li class="dropdown"><a href="#">{{ user.email }}</a>
                                             <ul>
+                                                <!--<li><a href="{% url 'profile' %}">Perfil</a></li>-->
                                                 <li><a href="{% url 'chat_room' %}">Chat de cooperativa</a></li>
                                                 <li><a href="{% url 'logout' %}">Cerrar Sesi&oacute;n</a></li>
                                             </ul>
@@ -208,7 +209,17 @@
                             <li><a href="{% url 'registrar_productor_administrador' %}">Activar correo</a></li>
                         </ul>
                     </li>
-                    <li><a href="#" data-toggle="modal" data-target="#login-modal">Login</a></li>
+                    {% if user.is_authenticated %}
+                        <li class="dropdown"><a href="#">{{ user.email }}</a>
+                            <ul>
+                                <!--<li><a href="{% url 'profile' %}">Perfil</a></li>-->
+                                <li><a href="{% url 'chat_room' %}">Chat de cooperativa</a></li>
+                                <li><a href="{% url 'logout' %}">Cerrar Sesi&oacute;n</a></li>
+                            </ul>
+                        </li>
+                    {% else %}
+                        <li><a href="#" data-toggle="modal" data-target="#login-modal">Login</a></li>
+                    {% endif %}
                     <!--<li class="dropdown"><a href="#">Blog</a>
                         <ul>
                             <li><a href="blog.html">Our Blog</a></li>
@@ -429,7 +440,7 @@
                     location.reload();
                 }
             },
-            complete: function( jqXHR, textStatus ) {
+            complete: function (jqXHR, textStatus) {
                 $("#loginButton").prop('disabled', false);
             }
         });
