+++ index/urls.py
@@ -23,4 +23,5 @@ urlpatterns = [
     url(r'^notfound$', views.notFound, name="not_found"),
     url(r'^login', views.login_view, name="login"),
     url(r'^logout', views.logout_view, name="logout"),
+    url(r'^profile', views.profile, name="profile"),
 ]
